Add files via upload
Exchange Rate Forecasting Challenge
Join the Exchange Rate Forecasting Challenge

Description
If you have any questions, please reach out to your ambassador Nasirudeen Raheem - Nasirudeen.

Nigerian economies rely on the dollar exchange rate to Naira, and the conversion rate has a strong influence on the Nigerian economy for things like buying goods, paying for services, doing international business, and so on. Over the years, there has been a constant fluctuation in the conversion rate, which causes a lot of dilemmas for Nigerian citizens. The challenge is open to statisticians, mathematicians, data scientists, machine learning enthusiasts, and all disciplines in the AI domain. The goal of the project is to build a time series or machine learning model that will accurately forecast what the naira rate against the dollar will be in the next 2 weeks, ranging from May 22, 2024, to June 4, 2024.
https://zindi.africa/competitions/exchange-rate-forecasting-challenge


rank : 2nd 
Name; Emmanuel Ebiendele
https://zindi.africa/competitions/exchange-rate-forecasting-challenge/leaderboard